+++ ZEA.Architecture.Pattern.Mediator.Abstractions/Interfaces/INotification.cs
@@ -1,0 +1,3 @@
+namespace ZEA.Architecture.Pattern.Mediator.Abstractions.Interfaces;
+
+public interface INotification;