+++ ZEA.Architecture.Pattern.Mediator.Abstractions/Interfaces/INotificationHandler.cs
@@ -1,0 +1,6 @@
+namespace ZEA.Architecture.Pattern.Mediator.Abstractions.Interfaces;
+
+public interface INotificationHandler<in TNotification>
+{
+	Task HandleAsync(TNotification notification, CancellationToken cancellationToken = default);
+}