+++ ZEA.Architecture.Pattern.Mediator.Abstractions/Responses/Conflict.cs
@@ -1,0 +1,4 @@
+﻿namespace ZEA.Architecture.Pattern.Mediator.Abstractions.Responses;
+
+[Obsolete("Use a type of ZEA.Architecture.Patterns.ADTs.Errors.LogicalErrors instead.")]
+public record Conflict(ErrorDetails Error);