+++ ZEA.Architecture.Pattern.Mediator.MediatrWrapper.Tests/Samples/SampleRequest.cs
@@ -1,0 +1,8 @@
+using ZEA.Architecture.Pattern.Mediator.Abstractions.Interfaces;
+
+namespace ZEA.Architecture.Pattern.Mediator.MediatrWrapper.Tests.Samples;
+
+public class SampleRequest : IRequest<string>
+{
+	public string Data { get; set; }
+}