+++ ZEA.Architecture.PubSub.Abstractions/ErrorDetails.cs
@@ -1,5 +1,0 @@
-﻿namespace ZEA.Architecture.PubSub.Abstractions;
-
-public sealed record ErrorDetails(
-	string? Code,
-	string? Message);