+++ ZEA.Architecture.PubSub.Abstractions/Interfaces/INotification.cs
@@ -1,3 +1,0 @@
-namespace ZEA.Architecture.PubSub.Abstractions.Interfaces;
-
-public interface INotification;