+++ ZEA.Architecture.PubSub.Abstractions/Interfaces/INotificationHandler.cs
@@ -1,6 +1,0 @@
-namespace ZEA.Architecture.PubSub.Abstractions.Interfaces;
-
-public interface INotificationHandler<in TNotification>
-{
-	Task HandleAsync(TNotification notification, CancellationToken cancellationToken = default);
-}