+++ ZEA.Architecture.PubSub.Abstractions/Interfaces/IRequest.cs
@@ -1,5 +1,0 @@
-namespace ZEA.Architecture.PubSub.Abstractions.Interfaces;
-
-public interface IRequest;
-
-public interface IRequest<out TResponse> : IRequest;