+++ ZEA.Architecture.PubSub.MediatrWrapper.Tests/Samples/SampleRequest.cs
@@ -1,8 +1,0 @@
-using ZEA.Architecture.PubSub.Abstractions.Interfaces;
-
-namespace ZEA.Architecture.PubSub.MediatrWrapper.Tests.Samples;
-
-public class SampleRequest : IRequest<string>
-{
-	public string Data { get; set; }
-}